+++ src/routes/store.js
@@ -2,3 +2,4 @@ import { writable } from "svelte/store";
 // @ts-ignore
 export const currencyStore = writable("USD ($)");
 export const updateRate = writable(15000);
+export const entryStore = writable(10);