+++ tailwind.config.js
@@ -25,6 +25,9 @@ module.exports = {
       fontFamily: {
         poppins: ["Poppins", "sans-serif"],
       },
+      screens: {
+        xs: "360px",
+      },
     },
   },
   plugins: [],